Fix some bit rot
Elasticsearch is now a few versions ahead (1.3.x vs 0.9.0) and it looks
like some versions of the pyelasticsearch package have disappeared (??).
Anyway, I tried to be as faithful to the original implementation with
these name changes, but it's possible there are semantic differences
that I didn't grok in my find-and-replace to remove the
then-deprecated-now-removed ElasticSearch query stuff.