Add basic Flask server for verifier
This adds a basic Flask server as a starting point for the rest of the
verifier. It's still in debug mode, and the API is definitely not how it
will end up being, but this will give me a good place to start
experimenting with deployment of microservices like this one.

Also this sneaks an .overcommit.yml in there to be compliant with
Brigade Best Practices. Also, this changes the location of the git
dependency to not use the SSH protocol to aid in deployment in scenarios
where SSH agent forwarding is impossible (building docker images)

Change-Id: Iddae59d818b327a90d4e808ca51053a745ad0a27