Add JSON-Schema validation
This is way overkill for this toy API, but it's fun and I want to learn
how JSON-Schema works so I'm doing it.

The schema.json file is the top-level fully-comprehensive
JSON-Hyper-Schema document which contains enough information for other
projects to construct API clients and validate inputs.

To generate a new version of that file, you'll need to:

  make schema.json

(You'll need to `gem install prmd`).

To add a new resource to the API, run the command:

  prmd init [resource_name] > schema/[resource_name].json

As with most Ruby "init" commands, you'll end up with way more than you
need, so help keep the schema clean by reducing that clutter. Running
`make schema.json` should cause it to join the schema.json file.

Versioning is to-be-determined.

Also, there is one glaring thing to fix: we re-load the schema from disk
for every API request. This will change, but I could use a little Python
guidance as to an idiomatic place to put it.

Change-Id: I11fcba3a0772877c86a9a0d5900fd9abef6d7697