Store and retrieve the full voterfile row from ES
Previously we passed '{ "_source": { "enabled": false } }' when we
created the ElasticSearch (ES) mapping. This results in a super-tiny
index because ES only keeps the information necessary to perform queries
-- not to return the original information back along with the query.

At votizen, they must have imported this data into a relational
database, but we can simplify the architecture for now by making the
persistence of this information a responsibility of ES.

Because we no longer have autoincrement voter IDs, we are now using the
raw IDs provided by TargetSmart ("OH-2098357083"). In light of that,
this commit also changes the type of the "_id" column to be a string.

Change-Id: Ia225db67b3e1b12fd3773038b1d96e57f17ff914