Add /health endpoint
As any service should have, this adds a health-check endpoint which
verifies that the ES cluster is reachable and is not in the "red" state,
meaning it can serve queries.

It's not perfect -- the cluster could be partitioned or the index empty,
but it gets us as close as feasible for a health-check.

Change-Id: I6002faf9bb8e6be3fdacf560d02b1eb00b9e56df