Add service deployment files
Whereas with Rails apps we used Capistrano for deployment, for these SOA
Marathon services we generate Docker images for each revision, push them
to our registry, and update the Marathon configuration with the name of
the new image.

This commit adds some files I've made to make this process easier.

To deploy, run `make deploy`. This will use `git archive` to create a
pristine copy of whatever commit you're on, and will build that commit
with Docker. Then, on successful build, it will update the service.json
file with the new SHA and upload the new version of that file to
Marathon.

There is a prompt in the middle to make sure the user wants to do the
deploy.

The Dockerfile and Makefile are carefully crafted to maximize reuse of
the Docker build cache (which is quite good). Comments in those files
explain the optimizations. With these optimizations, deploys take under
a minute.

Change-Id: I8fde3f78bcf86a1fbc56f03d09f08794cb2ca420