Move service config into deployment script
This is mostly because it seems counterintuitive to keep a JSON file
lying around that has the currently-deployed code revision in it -- such
a file is going to need to always be committed when it changes and this
will create a ton of noise.

Also this refactors the deploy script a bit to DRY up the marathon
connection logic (which is now slightly more complicated because it hits
the Marathon API via TLS).

Change-Id: Iaf82ccc5b2b895e4129b7422c291c4689f1b5f57