Add arcanist linting to the project
Summary:
This is kind of like Overcommit, but built into arc. I'm going to try
getting by without Overcommit in this repo. (Sorry @shane!)

This commit fixes a number of tiny style tweaks and moves some data
around so that the egregious line length limit violations are all in one
file that is ignored.

Change-Id: I2a37503074b6f15bf1d881009417ba7fe87d5b0f

Reviewers: noah, shane

Reviewed By: shane

Subscribers: shane

Differential Revision: https://phabricator.corp.causes.com/D541