Replace hardcoded ES hosts config with envvar
Summary:
Depending on environment, we're going to want to use different
elasticsearch config. (It appears that at Votizen they just tested
against their production cluster). But here, let's try to avoid that so
we don't accidentally drop an index we don't mean to.

To do that, I wrote a simple method which fetches the
ELASTICSEARCH_HOSTS environment variable and has a sensible fallback
(the default port on localhost) if it is not specified. This method
necessitated the rename of the file from `constants` to `config`.

Change-Id: I8003bba6b1b9ae673b7a9eaa339feae16b75d8f1

Reviewers: noah

Reviewed By: noah

Differential Revision: https://phabricator.corp.causes.com/D539