Index `effective_date` and `registration_date`
We want to be able to identify how relevant a voter record is based on
how old it is. We're grabbing these two fields because we believe that
it may help us determine relevant voter record.

Effective Date: Last verified occupancy date
Registration Date: Date voter registered to vote

(Although we're not quite sure what effective date is, if it's useless
we'll remove it).

Change-Id: I59f9832da5f53500bf5f4eb2d7c836e00b575bed
Reviewed-on: https://code.brigade.com/1764
Reviewed-by: Tom Dooner <tom.dooner@brigade.com>
Tested-by: Alex Ng <alex.ng@brigade.com>