Increase the S3 download link availability time
When we generate urls to download targetsmart from S3 we pass in a
"expires_in" parameter, which indicates how long the download link is
active for. Since downloading everything takes about ~8 hours I've
bumped the time the urls are valid for to 12 hours to ensure we're able
to download everything.

Change-Id: If7ab25cca7d6ac12f20cb9eaaa14fc9c0158f272
Reviewed-on: https://code.brigade.com/1818
Reviewed-by: Tom Dooner <tom.dooner@brigade.com>
Tested-by: Alex Ng <alex.ng@brigade.com>