Change indexed address fields
The 'tsmart_full_address' seems to be some kind of "smart" reduction of
all the addresses that TargetSmart knows about a person. This means that
it can deviate from the address where the person is actually
registered. Switching to 'vb.vf_reg_cass_address_full' means that we
always use the registered address (after it has been corrected by a
Coding Accuracy Support System, or CASS[1]).

We also added a field to the original document for the 'street number'
('vb.vf_reg_cass_street_num'), which we will use when we want to show an
'anonymized' address ("Baker Street" instead of "221B Baker Street").

[1]: https://en.wikipedia.org/wiki/Coding_Accuracy_Support_System

Change-Id: I3c0ec09871e80cdefd15351a53968af8f0208549
Reviewed-on: https://code.brigade.com/2869
Reviewed-by: Tom Dooner <tom.dooner@brigade.com>
Tested-by: Tom Dooner <tom.dooner@brigade.com>