Extract mapping of TargetSmart address data into a function
I just spent a lot of time loading context on the various addresses
available to us from TargetSmart, and I am reasonably confident that we
will need to make more adjustments in the future. By extracting this
into a function I have an obvious place to add a docstring describing
the format, and a long form comment about why we made certain decisions.

While doing this, I noticed that we had still neglected to update one of
the address columns in the previous commits, which would have resulted
in confusing addresses in the wrong states.

Change-Id: Ia276d915094c2999e819cff99a104c31c926f75f
Reviewed-on: https://code.brigade.com/2940
Reviewed-by: Alex Ng <alex.ng@brigade.com>
Reviewed-by: Jeff Ammons <jeff.ammons@brigade.com>
Tested-by: Noah Silas <noah.silas@brigade.com>