Install the index when it doesn't exist
When attempting to run the `index_all.py` script with a new cluster, it
would fail to create the mapping, leading us to attempt some console
shenanigans to actually install the index. By extracting the
`create_mapping` function, we can use it in both the "create" and
"destroy and recreate" cases.

Change-Id: Id8d0752e036727895fcbff8851c321ca289cf62c
Reviewed-on: https://code.brigade.com/3184
Reviewed-by: John Miller <john.miller@brigade.com>
Tested-by: Noah Silas <noah.silas@brigade.com>