Account for cases with bad data where 01-01-YYYY is returned
Some people, such as Miche's dad, have bad data which returns Jan 1st
instead of their actual birthday. Because we were requiring the date to
either be missing or exact match in the 'must' section of the query,
we're returning no results. This simply adds a case for matching
into this 'must'. It could slightly reduce our accuracy in some but
will allow us to match people in this state.

Change-Id: Iafff3ad6e21f0c3be067db301574494aaac468fe
Reviewed-on: https://code.brigade.com/4524
Reviewed-by: Noah Silas <noah.silas@brigade.com>
Tested-by: Alex Ng <alex.ng@brigade.com>