Replace S3 with HDFS in `list_files.py`
Now that we have a copy of the targetsmart data in HDFS, we should stop
incurring the bandwidth waste and cost of data transfer out of Amazon
whenever we want to reindex this data.

This uses the HDFS over HTTP application we set up for Hue. In the
future we will probably want to add security to it (as right now it is
unauthenticated), and so I added a bit of error checking into this
script so that it breaks with a good error message in that case.

Change-Id: Iea61b41af40f3e3e3d78dbc69dcf3a99c13ffa1c
Reviewed-on: https://code.brigade.com/6436
Reviewed-by: Shane da Silva <shane.dasilva@brigade.com>
Tested-by: Tom Dooner <tom.dooner@brigade.com>