Use timestamped index when indexing new records
It's no longer appropriate to drop the votizen_verifier index when we
want to reindex the data from scratch. Although we could probably update
many of the documents in-place, we would have no way to know which
records Targetsmart has deduplicated and thus removed. Also, we can make
the switchover from old -> new data much more atomic and fault-tolerant.

Thus, this configures the indexing process to use an index named like
votizen_verifier_20151013042000 based on when the indexing started.
Unfortunately, you cannot rename indexes, but we can create an alias of
votizen_verifier that points to votizen_verifier_20151013042000 with no
apparent downsides according to the documentation. This gives us
basically-atomic switchover and the ability to switch back if something
goes awry. The code to do this switchover will come later.

Change-Id: Iab3228b1c5ccd3ab454fc460baa6892c488426ec
Reviewed-on: https://code.brigade.com/6437
Reviewed-by: Shane da Silva <shane.dasilva@brigade.com>
Tested-by: Tom Dooner <tom.dooner@brigade.com>