Begin conversion of verifier to bli
The verifier was the first service deployed on Marathon, so it seems
fitting to be the first service deployed via bli on Marathon as well.

This gives the verifier a basic deploy.yml for development/testing
purposes, and I will add the marathon and chronos configuration later.

P.S. The Dockerfile changes should be ported into a brigade-python
docker base image, but for now I just put them here.

Change-Id: I4a3aac13470e00bef9cc61f0210803c34f55d94f
Reviewed-on: https://code.brigade.com/6683
Tested-by: Leeroy Jenkins <jenkins@brigade.com>
Reviewed-by: Shane da Silva <shane.dasilva@brigade.com>