Add script for targetsmart S3 -> HDFS
Perhaps this should live in pluribus. Either way, I'm committing it here
so it lives in source control.

Note that when I ran this the first time, it was on the gateway machine
itself, and for this version of the script, I added an SSH into the
picture. I think it should still work, famous last words.

Change-Id: I4efc7198ddde11a2804a52c34f8ce501691874f7
Reviewed-on: https://code.brigade.com/6539
Reviewed-by: Shane da Silva <shane.dasilva@brigade.com>
Tested-by: Tom Dooner <tom.dooner@brigade.com>