Make tests pass (with ES 1.7.3 and ES 2.0.0-beta2)
Since it looks like the verifier is now here to stay, it'll be good to
make sure we don't introduce any regressions by regularly running the
test suite that Votizen left us.

This change is a couple smaller changes:

1) Updating test fixtures with new columns that we have indexed
(apartment number columns) or relied upon in the response ('hits')

2) Updating the test suite in accordance to logic changes we've made in
birthdate processing. The test to ensure that unknown birthdate records
match higher than wrong birthdate records failed. I removed it because,
although we generally desire that behavior, it's unlikely that we'll end
up with such a situation because of the cleanliness of our TargetSmart
data (i.e. how records from a given state tend to all have the same
detail of birthdate information).

Also there was a change necessary that assumed the 5 year margin of
error that we reduced to 1 year in 78e10d9.

3) ES 2.0 compatibility. Two changes - "numeric_range" filter has been
deprecated in favor of "range". The range filter is cached by default
and uses the inverted index rather than fielddata, which I think will
result in less memory usage and marginal query performance difference
given that our main query has no aggregations.

Also, ES 2.0 removes the GET /[index]/_status endpoint in favor for just
GET /[index].

Change-Id: I2613578349fb4fed83889d61ec114218b78dede4
Reviewed-on: https://code.brigade.com/6626
Reviewed-by: John Miller <john.miller@brigade.com>
Reviewed-by: Shane da Silva <shane.dasilva@brigade.com>
Tested-by: Tom Dooner <tom.dooner@brigade.com>