Finish verifier config for bli-based-deploy
With this config, and minor adjustment to the entrypoint, `bli deploy`
is capable of deploying the verifier.

This is a bit different than in Ruby because virtualenv puts the full
path in the shims' shebangs, so we need to make sure the app is always
run out of /app. We can accomplish this by doing a simple check of path
and setting up a symlink if necessary.

Change-Id: Idf5445d3df5e0e6065e6a265ec7b18f77f0073f8
Reviewed-on: https://code.brigade.com/6724
Tested-by: Leeroy Jenkins <jenkins@brigade.com>
Reviewed-by: Shane da Silva <shane.dasilva@brigade.com>