Replace bin/deploy with pointer to install bli
This replaces the `bin/deploy` script with instructions on how to
install and deploy with bli. Also, this makes minor changes to the
deploy.yml so there is minimal diff when deploying, and so that the
config updates the existing marathon app now that I have verified it
works the same in staging.

Change-Id: I4c9a3c7c84f7e19defad4bcf4fe4bbd910edd538
Reviewed-on: https://code.brigade.com/6753
Tested-by: Leeroy Jenkins <jenkins@brigade.com>
Reviewed-by: Shane da Silva <shane.dasilva@brigade.com>