Rename "ideal_num_matches" param to "max_matches"
Thanks to Noah for suggesting a simpler way to accomplish the
"ideal_num_matches" goal of reducing the result set to a fixed size when
confident while also allowing clients to tell when the result quality is
not good:

By renaming this parameter to "max_matches", the intent is hopefully
clear that the verifier can (at its discretion) return up to that many
results. Keeping in line with the no-false-positives approach of the
verifier, we only return results if we are sure they are somewhat better
than the next batch of results.

Since no clients use this parameter yet, we should be fine to rename
this without worrying of backwards compatibility.

Change-Id: I72fcad7c962ec1948612f118f9a1244a066a60d0
Reviewed-on: https://code.brigade.zone/10009
Tested-by: Leeroy Jenkins <jenkins@brigade.com>
Reviewed-by: Shane da Silva <shane.dasilva@brigade.com>