Copy latest targetsmart files to targetsmart/latest
This alters scripts to copy and use the latest voter files to/from
targetsmart/latest. The current way of copying files to the targetsmart/
level is prone to introducing duplicate data if we are not careful to
remove old files from the directory before syncing. This ensures that
all files are isolated to clearly named directories.

Change-Id: If0b52304503fbab86226cb9e3d01e3f9f33e9b0a
Reviewed-on: https://code.brigade.zone/10740
Reviewed-by: Tom Dooner <tom.dooner@brigade.com>
Reviewed-by: Victor Kmita <victor.kmita@brigade.com>
Tested-by: Leeroy Jenkins <jenkins@brigade.com>