Use middle name as a booster
Fortunately we already index middle name. I gave the boost
factor a score of 3 becuase I want otherwise even scores
to have the matching middle name win. This adds the middle
name to the JSON schema as well. This will not currently
give a boost to people without middle names. We will have
to index the middle names as null if they are empty instead
of the empty string.

[SERV-233]

Change-Id: I3124d52d6a4fced49f31333c5ada0be50a828ee6
Reviewed-on: https://code.brigade.zone/10772
Tested-by: Leeroy Jenkins <jenkins@brigade.com>
Reviewed-by: Tom Dooner <tom.dooner@brigade.com>