Implement /v1/voters/<id>
This endpoint returns the _source of the voter record -- that is: every
column of the CSV/parquet voterfile we have indexed in elasticsearch.

This will be useful to pass data between the verifier and the Brigade
Rails application so the verifier can remain the canonical source of all
voterfile-related information and Brigade can sync from the verifier
when we add a new field.

Change-Id: I4ea15da73675cedcddaa63b126b9bc72c155e33e
Reviewed-on: https://code.brigade.zone/10961
Tested-by: Leeroy Jenkins <jenkins@brigade.com>
Reviewed-by: Shane da Silva <shane.dasilva@brigade.com>
Reviewed-by: Noah Silas <noah.silas@brigade.com>