Do not filter matches on dob day that have a month value of 1
Our new birthdate data seems to indicate that there are many
records that have a a dob day of 1, that are actually placeholders
for missing dob day. Previously we assumed missing dab day and
dob months were neceessarily always together. This does not seem
like it's the case with the new data.

Change-Id: I1fbdb914ca90d616b1372f9b140afb5fc8db6f31
Reviewed-on: https://code.brigade.zone/10977
Reviewed-by: Dave Kincade <dave.kincade@brigade.com>
Tested-by: Leeroy Jenkins <jenkins@brigade.com>