Always use latest `bli` gem version to test
Since we're iterating on bli quickly, and it doesn't take a long time to
install, let's just always install the most recent version of it before
each verifier test run.

This will give us quick feedback if there is a bli regression and allow
us to fix the problem by pushing new versions of bli with no code
changes required in the verifier.

Change-Id: Iee292f0a048d49c34b21c54dd729a27d85a11cdf
Reviewed-on: https://code.brigade.zone/11184
Tested-by: Leeroy Jenkins <jenkins@brigade.com>
Reviewed-by: Shane da Silva <shane.dasilva@brigade.com>