Update verifier docker image to be centos 7.2
The verifier was our first experiment with deploying via Docker, and we
have come a long way since then. We have upgraded our entire
infrastructure from 7.0 to 7.1 to 7.2 (mostly). We have moved `gosu`,
`git`, and `wget` into our base centos image because they are so useful.
And the base image's entrypoint script was modelled after the
functionality in this entrypoint, but is tested much better.

Change-Id: I5bbd0f7870bb4cdf697fb16fd94dfde76e84041a
Reviewed-on: https://code.brigade.zone/11249
Tested-by: Leeroy Jenkins <jenkins@brigade.com>
Reviewed-by: Victor Kmita <victor.kmita@brigade.com>
Reviewed-by: John Miller <john.miller@brigade.com>