Make virtualenv relocatable after `pip install`
One crucial detail I overlooked in the transition to our standard
brigade-centos image in b0e5f5 was that the custom entrypoint script in
this repo always ensured that the app was accessible from the /app
directory.

This turns out to be required for virtualenv: it hardcodes the absolute
path of the python interpreter into the shebangs for the executables in
`env/bin`. This broke when on mesos when the image build installed
virtualenv into /app but Mesos ran it in /mnt/mesos/sandbox.

Fortunately, virtualenv (somewhat experimentally?) supports a flag
`--relocatable` which makes all the shebangs relative. I tested this
locally with success, despite the somewhat foreboding warnings on the
virtualenv documentation regarding this feature:

  http://virtualenv.readthedocs.org/en/latest/userguide.html#making-environments-relocatable

Change-Id: I86a9d529fc452b80e08afd33c8d1ea4c3c558efa
Reviewed-on: https://code.brigade.zone/11407
Tested-by: Leeroy Jenkins <jenkins@brigade.com>
Reviewed-by: Shane da Silva <shane.dasilva@brigade.com>