Run Docker 1.12.1 in Dock container
We'd like to upgrade to Docker 1.12.1 across our entire infrastructure.

The version of Docker that runs within the Docker container is
independent of the one running that container, so we can update this
pretty easily.

While here I also upgraded docker-compose from 1.7.1 -> 1.8.0 so we
don't build extra images by doing this in separate commits.

[OPS-336]

Change-Id: Iba93d88078e6ffced5e2ed5bc526f674cda11b2b
Reviewed-on: https://code.brigade.zone/16847
Tested-by: Leeroy Jenkins <jenkins@brigade.com>
Reviewed-by: Shane da Silva <shane.dasilva@brigade.com>