Pin docker-api gem to 1.28.0
A recent change to this gem broke our build process. Since we don't have
time to investigate and come up with a deeper fix, I'm pinning the
version here so we aren't blocked.

This is yet another benefit of running these builds inside a container!

While here I disabled installing documentation for these gems since we
don't need it.

Change-Id: Idc57d5f3cb2ac2d012a96bcc92f13bc7ba026733
Reviewed-on: https://code.brigade.zone/18526
Tested-by: Leeroy Jenkins <jenkins@brigade.com>
Reviewed-by: Shane da Silva <shane.dasilva@brigade.com>