Use correct analyzer for compact fields
While we use the name_compact_analyzer for the compact fields,
we do not (effectively never) use it in query time. This means
people with names like "O'Brien" can only be matched if they
appear in TS as "O'Brien" and not "Obrien".

Fixed issue for all name fields and added a test.

Change-Id: Id80cab186b3fb29d0574f22be83fea04c9798389
Reviewed-on: https://code.brigade.zone/18949
Reviewed-by: Jim He <jim.he@brigade.com>
Tested-by: Leeroy Jenkins <jenkins@brigade.com>