Update Docker software to latest versions
This matches what we use across our infrastructure.

While here I cleaned up the Docker Compose configuration a bit. I also
added the --build flag to the `docker-compose up` command to ensure that
we always build the images (this won't result in any difference unless
the user changes the Dockerfile).

Change-Id: I1c5981125b1e389a6516ee8ae4bdd734f7fdfe82
Reviewed-on: https://code.brigade.zone/19487
Tested-by: Leeroy Jenkins <jenkins@brigade.com>
Reviewed-by: Shane da Silva <shane.dasilva@brigade.com>