Convert Verifier to deploy without BLI
This is a big change but the net effect is that we can now deploy the
verifier using just a Python script.

BLI was a good experiment but this is the only project that ended up
using it.

While here, I moved all Dock-related scripts in the `dock` directory,
and all Jenkins-related scripts in the `jenkins` directory.

Change-Id: Ifdaadef30f58fd82a65848de52c20e3bab621d45
Reviewed-on: https://code.brigade.zone/19525
Tested-by: Leeroy Jenkins <jenkins@brigade.com>
Reviewed-by: Shane da Silva <shane.dasilva@brigade.com>