Update deploy script
Update deploy script to match `auth`, `brigade-graphql`, and
`api-gateway` repositories. We're starting to converge on a script that
can be used to deploy any service.

While here, change the deployment to `chown` the source files so the
`nobody` user can read them without issue. This wasn't a problem but we
saw it become a problem in the `auth` repo, so preemptively change it.

Change-Id: I8641c185cdd590992dbf807a66c21131fdf7bf60
Reviewed-on: https://code.brigade.zone/19549
Reviewed-by: Shane da Silva <shane.dasilva@brigade.com>
Tested-by: Shane da Silva <shane.dasilva@brigade.com>