Explicitly specify which steps to perform
We want to be able to choose which steps of the build/push/deploy
process to run. Most often it will be all three, but over the course of
writing this deploy script I've had to selectively test certain subsets.

This will make it easier for me to test going forward.

This also exposes a flag that enables output of the deploy tag. This
will be useful when we want to coordinate multiple deploys of the same
image to all use the same tag (e.g. when we deploy a DB migration task
first, it should use the same image as the task that will ultimately run
the application).

Change-Id: I8c868a571bf6ff537024aa8eb5561c7677e1d86f
Reviewed-on: https://code.brigade.zone/19551
Tested-by: Leeroy Jenkins <jenkins@brigade.com>
Reviewed-by: Shane da Silva <shane.dasilva@brigade.com>