Remove UNIQUE constraint
This breaks when deploying to production as Marathon will be caught in a
"Waiting..." state while it tries to satisfy a constraint that can't be
satisfied (since there are more instances than hosts).

From a probability perspective, it's unlikely we're going to schedule
EVERY instance on a single host, so remove the constraint.

Change-Id: I6291985d2da2afb8ba2be4aac9cc7e3484e59f81
Reviewed-on: https://code.brigade.zone/19552
Tested-by: Leeroy Jenkins <jenkins@brigade.com>
Reviewed-by: Shane da Silva <shane.dasilva@brigade.com>