Add a score boost if city and state are included in the query
Context: When using Facebook for signout we only have city and state (not zip)
for the user so we need to added an approriate boost when searching occurs with
just name / city / state (and not zip code).

Added a simple test which includes matching on score, not sure
if this is too brittle.

Change-Id: I5df08ee163b4fbc8cf4ad7fe21d12778bf0bf8cc
Reviewed-on: https://code.brigade.zone/21709
Tested-by: Leeroy Jenkins <jenkins@brigade.com>
Reviewed-by: Dave Kincade <dave.kincade@brigade.com>