Bootstrapping the verifier service
PURPOSE:
We want the docker container verifier running within dock to be
built with the aritfacts within itself instead of just the
dock container (which is how it currently works)

FILE CHANGES:
Adding a bootstrap_dockerfile to the dock directory. This will
build the docker container verifier with the required artifacts.
Once the dock container comes up it uses the docker-compose file
which uses the bootstrap_dockerfile to build out the verifier
container

Change-Id: I53b38e5103b278ba0c140799c1bde480c61d7ab4
Reviewed-on: https://code.brigade.zone/21760
Reviewed-by: John Miller <john.miller@brigade.com>
Reviewed-by: Ahmad Wilson <ahmad.wilson@brigade.com>
Tested-by: Leeroy Jenkins <jenkins@brigade.com>