Adding healthchecks to docker compose for verifier service
The base compose has been updated to version file version 2.1 along
with the updated version of docker compose to file version 2.1. This
enables us to use the healthcheck feature in docker compose to 1.14.0.
The healthcheck SLA's (which can be adjusted) are shown below:

Healthchecks:
  Test port:  curl localhost:3000
  Interval:   1s
  Timeout:    1s
  retries:    120
  depends on Elasticsearch: service_healthy

Elasticsearch Healthchecks in the base compose
  test: curl localhost:9200

Change-Id: I10d68ad852f01843df61bf1148adb9e98e317c1c
Reviewed-on: https://code.brigade.zone/25021
Tested-by: Leeroy Jenkins <jenkins@brigade.com>
Reviewed-by: John Miller <john.miller@brigade.com>