Replace random score with random offset
When we do random matching (eg to come up with fake addresses), it seems
like we abuse ES quite heavily and as such can blow out memory.
Randomness is valuable to us still, but we can achieve something like it
via a random offset. It's not a great solution because the search space
is significantly smaller (50k here rather than the full space) and
because the offset is only deterministic per ES worker (or master?). As
far as I can tell, there are 3 sets of results we can get.

A more permanent fix could be to maintain a smaller index of samples by
state that we'd feel more comfortable iterating on as before. When we're
not on fire, that seems like a good move.

Change-Id: Icadc6354bf22973cc0cd9464c0b31d1dd7b9501c
Reviewed-on: https://code.brigade.zone/25455
Tested-by: Leeroy Jenkins <jenkins@brigade.com>
Reviewed-by: Anuj Das <anuj.das@brigade.com>