Rename console script to cluster_console to match service standard
Per a discussion between Ops/Infra and Backend, let's settle on a more
appropriate name/standard for the staging/prod environment console launching script
formerly known as "prodcmd".

"cluster_console" is expected to be a tunnel to a console within a node
preloaded with configuration specific to a particular environment for
this service.

Change-Id: Ifbd8433dcf565972a197431c47ca2e35c07a8741
Reviewed-on: https://code.brigade.zone/25690
Reviewed-by: Anuj Das <anuj.das@brigade.com>
Tested-by: Leeroy Jenkins <jenkins@brigade.com>